Fix crossed rows formula so new rows can be inserted
As it was previously, if a new row was inserted, the old formula would now skip over the new row when checking the end time of the previous cue.

New formula uses INDIRECT("F"&(ROW()-1)) to always check the row before, regardless of any insertions.

Also added an extra empty condition guard so that the formula would work in the first Cues row too, to make copying formulas straight up & down easier.